requirements.txt
Adafruit-Blinka==8.66.0
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-dht==4.0.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-typing==1.12.2
Adafruit-PlatformDetect==3.83.1
Adafruit-PureIO==1.1.11
awscrt==0.27.4
awsiotsdk==1.0.0.dev0
binho-host-adapter==0.1.6
colorzero==2.0
gpiozero==2.0.1
lgpio==0.2.2.0
pyftdi==0.57.1
pyserial==3.5
pyusb==1.3.1
rpi-ws281x==5.0.0
sysv-ipc==1.1.0
tinydb==4.8.2
typing_extensions==4.15.0
